+++ src/ui/widgets/window.rs
@@ -102,7 +102,7 @@ impl Window {
             context.window.position.x + context.window.size.x - style.title_height + 1.,
             context.window.position.y + 2.,
         );
-        let rect = Rect::new(pos.x, pos.y, size.x as f32, size.y as f32);
+        let rect = Rect::new(pos.x, pos.y, size.x, size.y);
         let (hovered, clicked) = context.register_click_intention(rect);
 
         context.window.painter.draw_element_background(
